chore(imports): Resolve circular dependencies
This moves the helpers and imports around to resolve the circular dependancy that rollup mentions.
